# Optimize validateLabels